feat(llm): add the analytics opt in prompt in the read only portfolio screen
