+++ apps/ledger-live-mobile/src/screens/AnalyticsOptInPrompt/variantB/Details.tsx
@@ -6,7 +6,7 @@ import Button from "~/components/Button";
 import { View, Container, Titles, Content, Bottom, ScrollableContainer } from "../Common";
 import headerPersonalized from "./illustrations/header_personalized.png";
 import { Image } from "react-native";
-import useAnalyticsOptInPrompt from "~/hooks/analyticsOptInPrompt/useAnalyticsOptInPromptVariantB";
+import useAnalyticsOptInPromptLogic from "~/hooks/analyticsOptInPrompt/useAnalyticsOptInPromptLogicVariantB";
 import { StackNavigatorProps } from "~/components/RootNavigator/types/helpers";
 import { AnalyticsOptInPromptNavigatorParamList } from "~/components/RootNavigator/types/AnalyticsOptInPromptNavigator";
 import { ScreenName } from "~/const";
@@ -25,7 +25,7 @@ function Details({ route }: Props) {
     clickOnRefusePersonalizedExperience,
     clickOnLearnMore,
     flow,
-  } = useAnalyticsOptInPrompt({ entryPoint });
+  } = useAnalyticsOptInPromptLogic({ entryPoint });
 
   const bulletPoints = [
     t("analyticsOptIn.variantB.details.bulletPoints.1"),
