+++ apps/ledger-live-mobile/src/screens/AnalyticsOptInPrompt/variantB/Main.tsx
@@ -6,7 +6,7 @@ import Button from "~/components/Button";
 import { View, Container, Titles, Content, Bottom, ScrollableContainer } from "../Common";
 import headerAnalytics from "./illustrations/header_analytics.png";
 import { Image } from "react-native";
-import useAnalyticsOptInPrompt from "~/hooks/analyticsOptInPrompt/useAnalyticsOptInPromptVariantB";
+import useAnalyticsOptInPromptLogic from "~/hooks/analyticsOptInPrompt/useAnalyticsOptInPromptLogicVariantB";
 import { StackNavigatorProps } from "~/components/RootNavigator/types/helpers";
 import { AnalyticsOptInPromptNavigatorParamList } from "~/components/RootNavigator/types/AnalyticsOptInPromptNavigator";
 import { ScreenName } from "~/const";
@@ -20,7 +20,7 @@ function Main({ route, navigation }: Props) {
   const { t } = useTranslation();
   const { entryPoint } = route.params;
   const { shouldWeTrack, clickOnAllowAnalytics, clickOnRefuseAnalytics, clickOnLearnMore, flow } =
-    useAnalyticsOptInPrompt({ entryPoint });
+    useAnalyticsOptInPromptLogic({ entryPoint });
 
   const shouldPreventBackNavigation = entryPoint === "Portfolio";
